adapt flutter drawer example
