+++ lib/main.dart
@@ -113,7 +113,6 @@ class _MyAppState extends State<MyApp> {
   Widget build(BuildContext context) {
     return OverlaySupport(
         child: MaterialApp(
-      title: 'Mosquito alert',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
         primarySwatch: Colors.orange,
